Rename topics domain to topic.
- ensure consistent naming
