+++ coral/src/app/features/topics/components/list/TopicCard.tsx
@@ -8,7 +8,7 @@ import {
   ExternalLinkButton,
 } from "@aivenio/design-system";
 import classes from "src/app/features/topics/components/list/TopicCard.module.css";
-import { Topic } from "src/domain/topics";
+import { Topic } from "src/domain/topic";
 import { createTopicOverviewLink } from "src/app/features/topics/utils/create-topic-overview-link";
 
 type TopicCardPros = Pick<
