+++ coral/src/app/features/topics/components/list/TopicList.test.tsx
@@ -1,6 +1,6 @@
 import { cleanup, screen, within } from "@testing-library/react";
 import { renderWithQueryClient } from "src/services/test-utils";
-import { createMockTopic } from "src/domain/topics/topic-test-helper";
+import { createMockTopic } from "src/domain/topic/topic-test-helper";
 import TopicList from "src/app/features/topics/components/list/TopicList";
 
 const mockedTopicNames = ["Name one", "Name two", "Name three", "Name four"];
