+++ coral/src/app/features/topics/components/list/TopicList.tsx
@@ -1,5 +1,5 @@
 import { Flexbox } from "@aivenio/design-system";
-import { Topic } from "src/domain/topics";
+import { Topic } from "src/domain/topic";
 import TopicCard from "src/app/features/topics/components/list/TopicCard";
 
 type TopicListProps = {
