+++ coral/src/app/features/topics/hooks/teams/useGetTeams.test.tsx
@@ -2,7 +2,7 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { renderHook, waitFor } from "@testing-library/react";
 import { ReactElement } from "react";
 import { server } from "src/services/api-mocks/server";
-import { mockGetTeams } from "src/domain/topics/topics-api.msw";
+import { mockGetTeams } from "src/domain/topic/topic-api.msw";
 import { useGetTeams } from "src/app/features/topics/hooks/teams/useGetTeams";
 
 const queryClient = new QueryClient({
