+++ coral/src/domain/topic/index.ts
@@ -1,0 +1,5 @@
+import { Topic, TopicEnv, TopicEnvDTO } from "src/domain/topic/topic-types";
+import { getTopics, getEnvs, getTeams } from "src/domain/topic/topic-api";
+
+export type { Topic, TopicEnvDTO };
+export { getTopics, getEnvs, getTeams, TopicEnv };
