--- coral/src/domain/topics/index.ts
@@ -1,5 +1,0 @@
-import { Topic, TopicEnv, TopicEnvDTO } from "src/domain/topics/topics-types";
-import { getTopics, getEnvs, getTeams } from "src/domain/topics/topics-api";
-
-export type { Topic, TopicEnvDTO };
-export { getTopics, getEnvs, getTeams, TopicEnv };
